Remove large auto-suggestions for Azure schemas
